+++ familyhelper-ui-view/packages/webapp-mobile/src/library/modules/simple/render/layout/navBar/Avatar.vue
@@ -25,7 +25,7 @@
 <script setup lang="ts">
 import vim from '@/vim'
 
-import type { LnpStore } from '@/store/modules/lnp.ts'
+import { type LnpStore } from '@/store/modules/lnp.ts'
 
 import { ref, onMounted, onUnmounted, computed, type ComputedRef } from 'vue'