+++ familyhelper-ui-view/packages/webapp-mobile/src/library/modules/simple/render/layout/sideBar/SideBar.vue
@@ -19,8 +19,8 @@
 <script setup lang="ts">
 import vim from '@/vim'
 
-import type { NavigationStore } from '@/store/modules/navigation.ts'
-import type { NavigationNodeInfo } from '@/navigation/types.ts'
+import { type NavigationStore } from '@/store/modules/navigation.ts'
+import { type NavigationNodeInfo } from '@/navigation/types.ts'
 
 import SidebarItem from './SidebarItem.vue'