+++ familyhelper-ui-view/packages/webapp-mobile/src/library/modules/simple/render/layout/sideBar/SidebarItem.vue
@@ -39,9 +39,9 @@
 <script setup lang="ts">
 import vim from '@/vim'
 
-import type { NavigationNodeInfo } from '@/navigation/types.ts'
-import type { NavigationStore } from '@/store/modules/navigation.ts'
-import type { LnpStore } from '@/store/modules/lnp.ts'
+import { type NavigationNodeInfo } from '@/navigation/types.ts'
+import { type NavigationStore } from '@/store/modules/navigation.ts'
+import { type LnpStore } from '@/store/modules/lnp.ts'
 
 import { computed, ref } from 'vue'