+++ familyhelper-ui-view/packages/webapp-mobile/src/library/modules/simple/render/login/Login.vue
@@ -17,7 +17,7 @@
 <script setup lang="ts">
 import vim from '@/vim'
 
-import type { LnpStore } from '@/store/modules/lnp.ts'
+import { type LnpStore } from '@/store/modules/lnp.ts'
 
 import { ref } from 'vue'